Renamed utility module to utilities to match convention
